compnent links for items
