+++ src/components/index.ts
@@ -1,2 +1,2 @@
 export { BootstrapClientInit } from "./BootstrapClientInit";
-export { Navbar } from "./Navbar";
+export { Navbar } from "./NavigationBar/index";
